datagen/pyproject.toml
[project]
name = "scrollstudy-datagen"
version = "0.1.0"
description = "Synthetic data generator for ScrollStudy experiment"
requires-python = ">=3.11"
dependencies = [
    "psycopg2-binary>=2.9",
    "pyyaml>=6.0",
    "numpy>=1.26",
    "scipy>=1.12",
    "python-dotenv>=1.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["datagen"]
